requirements.txt
fan-tools==2.12.1
httpx==0.16.1
pydantic==1.7.3
PyYAML==5.4.1
starlette==0.14.2
uvicorn==0.13.4
